Change the way available assets are calculated
I know, this is like a dejavu because in a past commit I 'improved'
this one, so why am I doing it again? Funny story:
I am kinda asshole sometimes so I reset a few commits and the code
of the 'improvement' dissappeared (for real), so I decided
to really improve this and as a result I created a new Jinja2 filter
that does all the calculations for us.

The commit where I thought I added this was 16fa312cf83f0d2126c7121779a7fbad4f8e1c26
Shame on me